gs: Increase ScheduleTimeLong to 500 ms
This increases the chance that packets arrive on time on gateways with a
high latency backhauls, like 3G.

This is a trade-off between delay in transmitting class C downlink
messages and making sure the packet arrives on time on the gateway. This
should serve as a reasonable default value for all gateways.