console: Fix form validation being blocked
When the form is submitted and native HTML form validation determines
there is an error within a field that is in a collapsed `FormCollapseSection`
component, an error is thrown due to the element not being "focusable."
The error then blocks the submit event so that Formik validation does not
run. This has been fixed by directly calling the Formik submit method while
preventing the default submit behavior.