ns: Use join-accept delay 2 for keeping join-request metadata
Join Servers may respond slower than the collection window that is used
for uplink messages. The join-request metadata for deduplication is now
valid as long as the join-request can be processed. This case is
different compared to uplink messages, where the collection window
should not be so long that retransmissions become part of it.