as: Abstract percentages and concentration parsing
These kind of of validation checks are so common and might appear even
more as the schema grows. In the same way that definitions are used in
the JSON schema for defining these types, here I've abstracted them to
separate functions to facilitate the maintenance down the road.

I've also added some test to increase coverage and explicitly test the
latest additions.