console: Clean event preview UI of normalized uplink messages
Remove FPort, data rate, SNR and RSSI from the preview of normalized
uplink messages. This info clutters the UI and is redundant here since
these values don't change between the normal uplink message and the
normalized uplink. The user can continue to view this info at a glance
in the correspondent uplink event or in the event details.